requirements.txt
streamlit==0.88.0
numpy=1.19.5
tensorflow=2.6.0
Pillow==7.2.0
